feat: create base Emoji class for ReactionEmoji and new GuildEmoji
