+++ src/client/actions/MessageReactionAdd.js
@@ -33,7 +33,7 @@ class MessageReactionAdd extends Action {
  * Emitted whenever a reaction is added to a message.
  * @event Client#messageReactionAdd
  * @param {MessageReaction} messageReaction The reaction object
- * @param {User} user The user that applied the emoji or reaction emoji
+ * @param {User} user The user that applied the guild or reaction emoji
  */
 
 module.exports = MessageReactionAdd;
